issue fix

+++ database/migrations/2022_06_12_103109_create_facilities_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('facilities', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->tinyInteger('status')->index()->default(1)->comment('1-active,0-deactivated');
+            $table->integer('status')->index()->default(1)->comment('1-active,0-deactivated');
             $table->string('created_by')->nullable();
             $table->string('updated_by')->nullable();
             $table->timestamps();
